Rename renaming libsodum module to ed25519
