+++ packages/crypto/src/ed25519.ts
