+++ packages/crypto/src/index.ts
@@ -1,9 +1,9 @@
 export { type Argon2idOptions, Argon2id, isArgon2idOptions } from "./argon2";
 export { Bip39, EnglishMnemonic } from "./bip39";
+export { Ed25519, Ed25519Keypair } from "./ed25519";
 export type { HashFunction } from "./hash";
 export { Hmac } from "./hmac";
 export { Keccak256, keccak256 } from "./keccak";
-export { Ed25519, Ed25519Keypair } from "./libsodium";
 export { Random } from "./random";
 export { Ripemd160, ripemd160 } from "./ripemd";
 export { type Secp256k1Keypair, Secp256k1 } from "./secp256k1";
